Fix burn dates
Presently, the first two days of BM week are annotated "Man Burns" and "Temple Burns."  I believe this is because the early arrival week also includes a Saturday.  Therefore, I've restricted `numSats` to only increment during the burn week.

I've made this change in the GitHub.com editor; I'm not a PHP developer, and I haven't tested this commit locally.